// sanity/types.ts
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export type STARTUP_QUERYResult = Array<{
  _id: string;
  title: string | null;
  slug: Slug | null;
  _createdAt: string;
  author: null;
  views: null;
  image: null;
  description: null;
  category: null;
  pitch: null;
} | {
  _id: string;
  title: null;
  slug: null;
  _createdAt: string;
  author: null;
  views: null;
  image: string | null;
  description: null;
  category: null;
  pitch: null;
} | {
  _id: string;
  title: string | null;
  slug: null;
  _createdAt: string;
  author: null;
  views: null;
  image: null;
  description: string | null;
  category: null;
  pitch: null;
} | {
  _id: string;
  title: string | null;
  slug: Slug | null;
  _createdAt: string;
  author: {
    _id: string;
    name: string | null;
    image: string | null;
    bio: string | null;
  } | null;
  views: number | null;
  image: string | null;
  description: string | null;
  category: string | null;
  pitch: string | null;
}>;